pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vintage-resale-mcp"
version = "0.1.0"
description = "MCP server for vintage resale — inventory search, pricing comps, listing management, and passive market signal collection"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Brandon", email = "brandon@example.com"},
]

dependencies = [
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
    "stripe>=7.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
reports = [
    "jinja2>=3.1",
    "weasyprint>=60.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
vintage-resale-mcp = "server:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
